# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OpenResidency API
  version: 0.1.0
  description: >
    Country-agnostic subnational Residency ID and SSO platform. Binds any national
    foundational ID to a W3C Verifiable Credential, issues cross-sector single
    sign-on, and records consent and a tamper-evident audit trail. The raw national
    ID is never stored or returned; only a tokenized subjectRef leaves the adapter.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /consent/resident/{residentId}:
    get:
      tags: [consent]
      summary: List consents a resident has granted
      security: [{ operatorKey: [] }, { operatorToken: [] }, { adminKey: [] }]
      parameters:
        - $ref: '#/components/parameters/ResidentId'
      responses:
        '200':
          description: Consent records
          content:
            application/json:
              schema:
                type: object
                properties:
                  residentId: { type: string }
                  consents:
                    type: array
                    items: { $ref: '#/components/schemas/ConsentRecord' }
  /consent/grant:
    post:
      tags: [consent]
      summary: Grant consent to a relying party and mint a signed receipt
      security: [{ operatorKey: [] }, { operatorToken: [] }, { adminKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsentGrantRequest'
      responses:
        '200':
          description: The consent record and its signed receipt (JWT)
          content:
            application/json:
              schema:
                type: object
                properties:
                  consent: { $ref: '#/components/schemas/ConsentRecord' }
                  receipt: { type: string, description: Signed consent receipt (EdDSA JWT) }
  /consent/{id}/revoke:
    post:
      tags: [consent]
      summary: Revoke a consent record
      security: [{ operatorKey: [] }, { operatorToken: [] }, { adminKey: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: The updated consent record
          content:
            application/json:
              schema:
                type: object
                properties:
                  consent: { $ref: '#/components/schemas/ConsentRecord' }
...
  /audit:
    get:
      tags: [audit]
      summary: Read the audit log (admin)
      security: [{ operatorKey: [] }, { operatorToken: [] }, { adminKey: [] }]
      parameters:
        - { name: limit, in: query, schema: { type: integer } }
        - { name: offset, in: query, schema: { type: integer } }
        - { name: target, in: query, schema: { type: string } }
      responses:
        '200':
          description: Audit events, newest first
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
                  events:
                    type: array
                    items: { $ref: '#/components/schemas/AuditEvent' }
        '401': { description: Missing or invalid admin key }
  /audit/verify:
    get:
      tags: [audit]
      summary: Verify the audit hash chain is intact (admin)
      security: [{ operatorKey: [] }, { operatorToken: [] }, { adminKey: [] }]
      responses:
        '200':
          description: Chain integrity result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  length: { type: integer }
                  brokenAtSeq: { type: integer, nullable: true }
...
  /offline/qr:
    post:
      tags: [offline]
      summary: Render a residency credential as a scannable QR (SVG)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                credential: { type: string }
                residentId: { type: string }
      responses:
        '200':
          description: SVG QR
          content:
            image/svg+xml:
              schema: { type: string }
  /offline/ussd:
    post:
      tags: [offline]
      summary: USSD webhook for feature phones
      description: >
        Called by the USSD aggregator only. The handler trusts the caller's word for
        `phoneNumber`, so it requires the shared gateway secret.
      security: [{ ussdSecret: [] }]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                text: { type: string }
                phoneNumber: { type: string }
      responses:
        '200':
          description: USSD CON/END response
          content:
            text/plain:
              schema: { type: string }
...
components:
  parameters:
    ResidentId:
      name: residentId
      in: path
      required: true
      schema: { type: string, example: KT-GT1F-75WJ-6 }
  securitySchemes:
    operatorKey:
      type: apiKey
      in: header
      name: x-operator-key
      description: >
        Per-operator API key (ork_...), minted at POST /operator/keys. Identifies which
        operator is calling, carries only that operator's roles, expires, and rotates with
        an overlap window. This is the credential to use.
    operatorToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        An operator access token -- from the configured identity provider when
        operatorAuth.mode is oidc, or from POST /operator/login when it is local.
    adminKey:
      type: apiKey
      in: header
      name: x-admin-key
      description: >
        DEPRECATED legacy shared key. Accepted only where operatorAuth.mode is sharedKey.
        Carries no operator identity, no roles, and cannot be rotated without a restart.
    ussdSecret:
      type: apiKey
      in: header
      name: x-ussd-secret
      description: >
        Shared secret the USSD aggregator returns on webhook callbacks. Set as
        USSD_GATEWAY_SECRET; the route fails closed when it is unset.
    walletToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        The OpenID4VCI access token obtained from /openid4vci/token. Scoped to a single
        credential offer: it authorizes only the credential configurations that offer
        named, for the resident that offer was created for.
  schemas:
...
    ConsentGrantRequest:
      type: object
      required: [residentId, relyingParty, purpose, scopes]
      properties:
        residentId: { type: string }
        subjectRef: { type: string }
        relyingParty: { type: string, example: health }
        relyingPartyName: { type: string }
        purpose: { type: string }
        scopes:
          type: array
          items: { type: string }
        validityDays: { type: integer }
    ConsentRecord:
      type: object
      properties:
        id: { type: string }
        residentId: { type: string }
        relyingParty: { type: string }
        purpose: { type: string }
        scopes:
          type: array
          items: { type: string }
        status: { type: string, enum: [active, revoked, expired] }
        grantedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time, nullable: true }
        revokedAt: { type: string, format: date-time, nullable: true }
        receiptId: { type: string }
    AuditEvent:
      type: object
      properties:
        seq: { type: integer }
        id: { type: string }
        timestamp: { type: string, format: date-time }
        action: { type: string }
        actor: { type: string }
        target: { type: string, nullable: true }
        countryCode: { type: string, nullable: true }
        outcome: { type: string, enum: [success, failure] }
        metadata: { type: object, additionalProperties: true, nullable: true }
        prevHash: { type: string }
        hash: { type: string }